+++ src/gameobjects/glyphmap.spec.ts
@@ -1,3 +1,5 @@
+import { phaserWebGLRendererMockAdapter } from '../../test/mocks/phaser-webgl-renderer-mock-adapter';
+
 import { GlyphPlugin } from '../plugin';
 import { Font, GlyphLike } from '../shared';
 
@@ -397,7 +399,7 @@ describe('Glyphmap', () => {
 
     const actual = (() => {
       const glyphmap = new Glyphmap(scene);
-      const spy = jest.spyOn(glyphmap, 'glyphPluginDestroyEventListener' as never);
+      const spy = jest.spyOn(glyphmap, 'glyphPluginDestroyEventListener');
       glyphmap.glyphPlugin = input;
       input.destroy();
       glyphmap.destroy();
@@ -414,7 +416,7 @@ describe('Glyphmap', () => {
 
     const actual = (() => {
       const glyphmap = new Glyphmap(scene);
-      const spy = jest.spyOn(glyphmap, 'refresh' as never);
+      const spy = jest.spyOn(glyphmap, 'refresh');
       glyphmap.glyphPlugin.measurementCh = input[0];
       glyphmap.glyphPlugin.advancedTextMetrics = input[1];
       glyphmap.glyphPlugin = input[2];
@@ -714,51 +716,7 @@ describe('Glyphmap', () => {
   });
 
   describe('WEBGL_RENDERER', () => {
-    beforeAll(() => {
-      const ctx = game.canvas.getContext('webgl');
-      const proto = window.WebGLRenderingContext.prototype;
-
-      ctx.isContextLost = () => false;
-      ctx.getSupportedExtensions = proto.getSupportedExtensions.bind(ctx);
-      ctx.getParameter = () => 16;
-      ctx.getExtension = proto.getExtension.bind(ctx);
-      ctx.disable = proto.disable.bind(ctx);
-      ctx.enable = proto.enable.bind(ctx);
-      ctx.clearColor = proto.clearColor.bind(ctx);
-      ctx.activeTexture = proto.activeTexture.bind(ctx);
-      ctx.blendEquation = proto.blendEquation.bind(ctx);
-      ctx.blendFunc = proto.blendFunc.bind(ctx);
-      ctx.createTexture = proto.createTexture.bind(ctx);
-      ctx.deleteTexture = proto.deleteTexture.bind(ctx);
-      ctx.bindTexture = proto.bindTexture.bind(ctx);
-      ctx.texImage2D = proto.texImage2D.bind(ctx);
-      ctx.bindBuffer = proto.bindBuffer.bind(ctx);
-      ctx.createProgram = proto.createProgram.bind(ctx);
-      ctx.createShader = proto.createShader.bind(ctx);
-      ctx.shaderSource = proto.shaderSource.bind(ctx);
-      ctx.compileShader = proto.compileShader.bind(ctx);
-      ctx.getShaderParameter = proto.getShaderParameter.bind(ctx);
-      ctx.attachShader = proto.attachShader.bind(ctx);
-      ctx.linkProgram = proto.linkProgram.bind(ctx);
-      ctx.getProgramParameter = proto.getProgramParameter.bind(ctx);
-      ctx.useProgram = proto.useProgram.bind(ctx);
-      ctx.createBuffer = proto.createBuffer.bind(ctx);
-      ctx.bufferData = proto.bufferData.bind(ctx);
-      ctx.getAttribLocation = proto.getAttribLocation.bind(ctx);
-      ctx.disableVertexAttribArray = proto.disableVertexAttribArray.bind(ctx);
-      ctx.deleteBuffer = proto.deleteBuffer.bind(ctx);
-      ctx.texParameteri = proto.texParameteri.bind(ctx);
-      ctx.pixelStorei = proto.pixelStorei.bind(ctx);
-      ctx.createFramebuffer = proto.createFramebuffer.bind(ctx);
-      ctx.bindFramebuffer = proto.bindFramebuffer.bind(ctx);
-      ctx.viewport = proto.viewport.bind(ctx);
-      ctx.framebufferTexture2D = proto.framebufferTexture2D.bind(ctx);
-      ctx.checkFramebufferStatus = () => undefined;
-      ctx.generateMipmap = proto.generateMipmap.bind(ctx);
-      ctx.scissor = proto.scissor.bind(ctx);
-      ctx.bufferSubData = proto.bufferSubData.bind(ctx);
-      ctx.drawArrays = proto.drawArrays.bind(ctx);
-    });
+    beforeAll(() => phaserWebGLRendererMockAdapter(game));
 
     it('renders an empty glyphmap', () => {
       const input = [new Phaser.Renderer.WebGL.WebGLRenderer(game), new Glyphmap(scene), scene.cameras.main] as const;